+++ files/private-chef-cookbooks/private-chef/recipes/opscode-solr4.rb
@@ -90,7 +90,7 @@ node.default['private_chef']['opscode-solr4']['command'] =  "java -Xmx#{node['pr
 # Compute some sane JVM tunings. The user can still override these computed
 # defaults using /etc/opscode/private-chef.rb
 solr_mem = if node['private_chef']['opscode-solr4']['heap_size']
-              node['private_chef']['opscode-solr4']['heap_size']
+             OmnibusHelper.parse_mem_to_mb(node['private_chef']['opscode-solr4']['heap_size'])
            else
              node[:memory][:total] =~ /^(\d+)kB/
              memory_total_in_mb = $1.to_i / 1024
@@ -100,7 +100,7 @@ solr_mem = if node['private_chef']['opscode-solr4']['heap_size']
              [(memory_total_in_mb / 4), 1024].min
            end
 new_size =  if node['private_chef']['opscode-solr4']['new_size']
-              node['private_chef']['opscode-solr4']['new_size']
+              OmnibusHelper.parse_mem_to_mb(node['private_chef']['opscode-solr4']['new_size'])
             else
               [(solr_mem / 16), 32].max
             end